♻️ remove home classic container
