+++ src/packages/@ncigdc/modern_components/HumanBody/index.js
@@ -1,0 +1,3 @@
+import Component from './HumanBody';
+import createRenderer from './HumanBody.relay';
+export default createRenderer(Component);
