+++ src/packages/@ncigdc/modern_components/PortalSummary/index.js
@@ -1,0 +1,3 @@
+import Component from './PortalSummary';
+import createRenderer from './PortalSummary.relay';
+export default createRenderer(Component);
